chore: reduce pkgs
