+++ .npmrc
@@ -1,1 +1,2 @@
 strict-peer-dependencies=false
+package-manager-strict=false
