+++ examples/solid-ts/package.json
@@ -47,7 +47,6 @@
     "@zag-js/file-upload": "workspace:*",
     "@zag-js/file-utils": "workspace:*",
     "@zag-js/floating-panel": "workspace:*",
-    "@zag-js/focus-scope": "workspace:*",
     "@zag-js/focus-visible": "workspace:*",
     "@zag-js/form-utils": "workspace:*",
     "@zag-js/hover-card": "workspace:*",
@@ -76,7 +75,6 @@
     "@zag-js/splitter": "workspace:*",
     "@zag-js/store": "workspace:*",
     "@zag-js/switch": "workspace:*",
-    "@zag-js/tabbable": "workspace:*",
     "@zag-js/tabs": "workspace:*",
     "@zag-js/tags-input": "workspace:*",
     "@zag-js/text-selection": "workspace:*",
