+++ examples/svelte-ts/package.json
@@ -39,7 +39,6 @@
     "@zag-js/file-upload": "workspace:*",
     "@zag-js/file-utils": "workspace:*",
     "@zag-js/floating-panel": "workspace:*",
-    "@zag-js/focus-scope": "workspace:*",
     "@zag-js/focus-visible": "workspace:*",
     "@zag-js/form-utils": "workspace:*",
     "@zag-js/hover-card": "workspace:*",
@@ -68,7 +67,6 @@
     "@zag-js/store": "workspace:*",
     "@zag-js/svelte": "workspace:*",
     "@zag-js/switch": "workspace:*",
-    "@zag-js/tabbable": "workspace:*",
     "@zag-js/tabs": "workspace:*",
     "@zag-js/tags-input": "workspace:*",
     "@zag-js/text-selection": "workspace:*",
