+++ package.json
@@ -108,7 +108,7 @@
   "engines": {
     "node": ">=18.0.0"
   },
-  "packageManager": "pnpm@9.1.0",
+  "packageManager": "pnpm@9.1.1",
   "devDependencies": {
     "cross-env": "^7.0.3"
   }
