+++ packages/machines/color-picker/src/color-picker.machine.ts
@@ -2,10 +2,9 @@ import { parseColor, type Color } from "@zag-js/color-utils"
 import { createMachine, guards } from "@zag-js/core"
 import { trackDismissableElement } from "@zag-js/dismissable"
 import { trackPointerMove } from "@zag-js/dom-event"
-import { raf } from "@zag-js/dom-query"
+import { getInitialFocus, raf } from "@zag-js/dom-query"
 import { dispatchInputValueEvent, trackFormControl } from "@zag-js/form-utils"
 import { getPlacement } from "@zag-js/popper"
-import { getInitialFocus } from "@zag-js/tabbable"
 import { disableTextSelection } from "@zag-js/text-selection"
 import { compact, tryCatch } from "@zag-js/utils"
 import { dom } from "./color-picker.dom"
