+++ packages/machines/dialog/package.json
@@ -31,7 +31,6 @@
     "@zag-js/aria-hidden": "workspace:*",
     "@zag-js/core": "workspace:*",
     "@zag-js/dom-query": "workspace:*",
-    "@zag-js/tabbable": "workspace:*",
     "@zag-js/utils": "workspace:*",
     "@zag-js/dismissable": "workspace:*",
     "@zag-js/remove-scroll": "workspace:*",
