+++ packages/machines/menu/package.json
@@ -35,7 +35,6 @@
     "@zag-js/utils": "workspace:*",
     "@zag-js/dismissable": "workspace:*",
     "@zag-js/popper": "workspace:*",
-    "@zag-js/tabbable": "workspace:*",
     "@zag-js/types": "workspace:*"
   },
   "devDependencies": {
