+++ packages/machines/popover/src/popover.machine.ts
@@ -1,10 +1,9 @@
 import { ariaHidden } from "@zag-js/aria-hidden"
 import { createMachine } from "@zag-js/core"
 import { trackDismissableElement } from "@zag-js/dismissable"
-import { nextTick, raf } from "@zag-js/dom-query"
+import { getInitialFocus, nextTick, proxyTabFocus, raf } from "@zag-js/dom-query"
 import { getPlacement } from "@zag-js/popper"
 import { preventBodyScroll } from "@zag-js/remove-scroll"
-import { getInitialFocus, proxyTabFocus } from "@zag-js/tabbable"
 import { compact } from "@zag-js/utils"
 import { createFocusTrap, type FocusTrap } from "focus-trap"
 import { dom } from "./popover.dom"
