+++ packages/machines/select/package.json
@@ -43,7 +43,6 @@
     "@zag-js/form-utils": "workspace:*",
     "@zag-js/utils": "workspace:*",
     "@zag-js/dismissable": "workspace:*",
-    "@zag-js/tabbable": "workspace:*",
     "@zag-js/types": "workspace:*"
   },
   "devDependencies": {
