+++ packages/machines/select/src/select.machine.ts
@@ -1,9 +1,8 @@
 import { createMachine, guards } from "@zag-js/core"
 import { trackDismissableElement } from "@zag-js/dismissable"
-import { getByTypeahead, observeAttributes, raf, scrollIntoView } from "@zag-js/dom-query"
+import { getByTypeahead, getInitialFocus, observeAttributes, raf, scrollIntoView } from "@zag-js/dom-query"
 import { trackFormControl } from "@zag-js/form-utils"
 import { getPlacement } from "@zag-js/popper"
-import { getInitialFocus } from "@zag-js/tabbable"
 import { addOrRemove, compact, isEqual } from "@zag-js/utils"
 import { collection } from "./select.collection"
 import { dom } from "./select.dom"
