+++ packages/machines/tabs/package.json
@@ -31,7 +31,6 @@
     "@zag-js/dom-query": "workspace:*",
     "@zag-js/dom-event": "workspace:*",
     "@zag-js/element-rect": "workspace:*",
-    "@zag-js/tabbable": "workspace:*",
     "@zag-js/utils": "workspace:*",
     "@zag-js/core": "workspace:*",
     "@zag-js/types": "workspace:*"
