+++ packages/machines/tabs/src/tabs.machine.ts
@@ -1,8 +1,7 @@
 import { createMachine, guards } from "@zag-js/core"
 import { clickIfLink } from "@zag-js/dom-event"
-import { nextTick, raf } from "@zag-js/dom-query"
+import { nextTick, raf, getFocusables } from "@zag-js/dom-query"
 import { trackElementRect } from "@zag-js/element-rect"
-import { getFocusables } from "@zag-js/tabbable"
 import { compact, isEqual } from "@zag-js/utils"
 import { dom } from "./tabs.dom"
 import type { MachineContext, MachineState, UserDefinedContext } from "./tabs.types"
