+++ packages/utilities/dom-query/src/index.ts
@@ -9,6 +9,7 @@ export * from "./get-by-typeahead"
 export * from "./get-computed-style"
 export * from "./get-parent-node"
 export * from "./get-scroll-position"
+export * from "./initial-focus"
 export * from "./is"
 export * from "./is-editable-element"
 export * from "./is-hidden-element"
@@ -17,8 +18,10 @@ export * from "./observe-attributes"
 export * from "./observe-children"
 export * from "./overflow"
 export * from "./platform"
+export * from "./proxy-tab-focus"
 export * from "./query"
 export * from "./raf"
 export * from "./scope"
 export * from "./scroll-into-view"
+export * from "./tabbable"
 export * from "./visually-hidden"
