--- packages/utilities/focus-scope/README.md
@@ -1,19 +1,0 @@
-# @zag-js/focus-scope
-
-Manage and trap focus within an element
-
-## Installation
-
-```sh
-yarn add @zag-js/focus-scope
-# or
-npm i @zag-js/focus-scope
-```
-
-## Contribution
-
-Yes please! See the [contributing guidelines](https://github.com/chakra-ui/zag/blob/main/CONTRIBUTING.md) for details.
-
-## Licence
-
-This project is licensed under the terms of the [MIT license](https://github.com/chakra-ui/zag/blob/main/LICENSE).
