--- packages/utilities/focus-scope/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@zag-js/focus-scope",
-  "version": "0.49.0",
-  "description": "Manage and trap focus within an element",
-  "keywords": [
-    "js",
-    "utils",
-    "focus",
-    "trap",
-    "focus-scope",
-    "trap-focus"
-  ],
-  "author": "Segun Adebayo <sage@adebayosegun.com>",
-  "homepage": "https://github.com/chakra-ui/zag#readme",
-  "license": "MIT",
-  "repository": "https://github.com/chakra-ui/zag/tree/main/packages/utilities/focus-scope",
-  "sideEffects": false,
-  "files": [
-    "dist",
-    "src"
-  ],
-  "scripts": {
-    "build": "tsup",
-    "test": "jest --config ../../../jest.config.js --rootDir tests",
-    "lint": "eslint src --ext .ts,.tsx",
-    "test-ci": "pnpm test --ci --runInBand -u",
-    "test-watch": "pnpm test --watchAll",
-    "prepack": "clean-package",
-    "postpack": "clean-package restore"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "@zag-js/dom-query": "workspace:*",
-    "@zag-js/tabbable": "workspace:*",
-    "@zag-js/dom-event": "workspace:*",
-    "@zag-js/utils": "workspace:*"
-  },
-  "devDependencies": {
-    "clean-package": "2.2.0"
-  },
-  "bugs": {
-    "url": "https://github.com/chakra-ui/zag/issues"
-  },
-  "clean-package": "../../../clean-package.config.json",
-  "main": "src/index.ts"
-}
