+++ packages/utilities/interact-outside/package.json
@@ -28,7 +28,6 @@
   "dependencies": {
     "@zag-js/dom-query": "workspace:*",
     "@zag-js/dom-event": "workspace:*",
-    "@zag-js/tabbable": "workspace:*",
     "@zag-js/utils": "workspace:*"
   },
   "devDependencies": {
