--- packages/utilities/tabbable/README.md
@@ -1,19 +1,0 @@
-# @zag-js/tabbable
-
-Small utility that returns an array of all\* tabbable DOM nodes within a containing node.
-
-## Installation
-
-```sh
-yarn add @zag-js/tabbable
-# or
-npm i @zag-js/tabbable
-```
-
-## Contribution
-
-Yes please! See the [contributing guidelines](https://github.com/chakra-ui/zag/blob/main/CONTRIBUTING.md) for details.
-
-## Licence
-
-This project is licensed under the terms of the [MIT license](https://github.com/chakra-ui/zag/blob/main/LICENSE).
