--- packages/utilities/tabbable/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@zag-js/tabbable",
-  "version": "0.49.0",
-  "description": "Small utility that returns an array of all* tabbable DOM nodes within a containing node.",
-  "keywords": [
-    "js",
-    "utils",
-    "tabbable"
-  ],
-  "author": "Segun Adebayo <sage@adebayosegun.com>",
-  "homepage": "https://github.com/chakra-ui/zag#readme",
-  "license": "MIT",
-  "main": "src/index.ts",
-  "repository": "https://github.com/chakra-ui/zag/tree/main/packages/utilities/tabbable",
-  "sideEffects": false,
-  "files": [
-    "dist",
-    "src"
-  ],
-  "scripts": {
-    "build": "tsup",
-    "lint": "eslint src --ext .ts,.tsx",
-    "typecheck": "tsc --noEmit",
-    "prepack": "clean-package",
-    "postpack": "clean-package restore"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "bugs": {
-    "url": "https://github.com/chakra-ui/zag/issues"
-  },
-  "clean-package": "../../../clean-package.config.json",
-  "devDependencies": {
-    "clean-package": "2.2.0"
-  },
-  "dependencies": {
-    "@zag-js/dom-query": "workspace:*"
-  }
-}
