--- packages/utilities/tabbable/src/index.ts
@@ -1,11 +1,0 @@
-export { getFirstFocusable, getFocusables, isFocusable } from "./focusable"
-export { getInitialFocus, isValidTabEvent } from "./initial-focus"
-export { proxyTabFocus } from "./proxy-tab-focus"
-export {
-  getFirstTabbable,
-  getLastTabbable,
-  getNextTabbable,
-  getTabbableEdges,
-  getTabbables,
-  isTabbable,
-} from "./tabbable"
