--- packages/utilities/tabbable/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "include": ["src"],
-  "compilerOptions": {
-    "tsBuildInfoFile": "node_modules/.cache/.tsbuildinfo"
-  }
-}
